+++ src/framework/common/service-worker-updater.js
@@ -8,30 +8,67 @@ import { useEffect, useState } from 'react';
 let updateAvailable = false;
 const updateListeners = new Set();
 
+// https://whatwebcando.today/articles/handling-service-worker-updates/
 export function updateServiceWorker() {
 
-  if (process.env.SIDE === 'client') {
-    try {
-      const runtime = require('offline-plugin/runtime');
+  if (process.env.SIDE !== 'client') {
+    return;
+  }
 
-      runtime.install({
-        onUpdateReady: () => {
-          runtime.applyUpdate();
-        },
-        onUpdated: () => {
-          updateAvailable = true;
-          for (const listener of updateListeners) {
-            listener(updateAvailable);
+  if (!('serviceWorker' in navigator)) {
+    return;
+  }
+
+  navigator.serviceWorker.register('/service-worker.js')
+    .then(registration => {
+      console.info('[SW] Registered');
+
+      if (registration.waiting) {
+        declareRestartRequired();
+      }
+
+      registration.addEventListener('updatefound', () => {
+        // If updatefound is fired, it means that there's
+        // a new service worker being installed.
+        const installingWorker = registration.installing;
+        if (!installingWorker) {
+          return;
+        }
+
+        console.info('[SW] Update found');
+
+        installingWorker.addEventListener('statechange', () => {
+          if (registration.waiting) {
+            declareRestartRequired();
           }
-        },
+        });
       });
-    } catch (e) {
-      console.error('Service Worker update failed');
-      console.error(e);
+    })
+    .catch(registrationError => {
+      console.error('[SW] registration failed', registrationError);
+    });
+
+  let refreshing = false;
+  navigator.serviceWorker.addEventListener('controllerchange', () => {
+    if (!refreshing) {
+      refreshing = true; // prevent calling reload twice
+      window.location.reload();
     }
+  });
+}
+
+function declareRestartRequired() {
+  updateAvailable = true;
+  for (const listener of updateListeners) {
+    listener(true);
   }
 }
 
+/**
+ * React Hook to detect whether a service-worker is awaiting activation.
+ *
+ * @returns {boolean} Whether a service-worker is awaiting activation.
+ */
 export function useRestartRequired() {
 
   const [hasUpdate, setHasUpdate] = useState(updateAvailable);
@@ -49,4 +86,19 @@ export function useRestartRequired() {
   }, []);
 
   return hasUpdate;
+}
+
+/**
+ * Cause the waiting service worker to activate, becoming the controlling SW,
+ *  which in turn will cause the page to refresh.
+ *
+ * @returns {Promise<void>}
+ */
+export async function updateSwAndRestart() {
+  const registration = await navigator.serviceWorker.ready;
+
+  if (registration.waiting) {
+    // let waiting Service Worker know it should became active
+    registration.waiting.postMessage('SKIP_WAITING');
+  }
 }