+++ src/internals/webpack/features/optimize.js
@@ -1,7 +1,7 @@
 // @flow
 
 import TerserPlugin from 'terser-webpack-plugin';
-import OptimizeCSSAssetsPlugin from 'optimize-css-assets-webpack-plugin';
+import CssMinimizerPlugin from 'css-minimizer-webpack-plugin';
 
 import BaseFeature from '../BaseFeature';
 import type WebpackConfigBuilder from '../WebpackConfigBuilder';
@@ -32,7 +32,7 @@ export default class OptimizeFeature extends BaseFeature {
     config.injectRules({
       test: BaseFeature.FILE_TYPE_IMG,
       loader: 'image-webpack-loader',
-      query: {
+      options: {
         bypassOnDebug: true,
         mozjpeg: {
           progressive: true,
@@ -78,20 +78,14 @@ export default class OptimizeFeature extends BaseFeature {
       },
       optimization: {
         removeAvailableModules: true,
-        splitChunks: {
-          // don't generate names for long term caching
-          // https://medium.com/webpack/webpack-4-code-splitting-chunk-graph-and-the-splitchunks-optimization-be739a861366
-          name: false,
-        },
         minimize: true,
         minimizer: [
           // minimize CSS
-          new OptimizeCSSAssetsPlugin({}),
+          new CssMinimizerPlugin(),
 
           // minimize JS
           new TerserPlugin({
             parallel: true,
-            cache: true,
 
             // preserve LICENSE comments (*!, /**!, @preserve or @license) for legal stuff but extract them
             // to their own file to reduce bundle size.