+++ src/internals/webpack/features/react-intl.js
@@ -38,10 +38,10 @@ export default class ReactIntlFeature extends BaseFeature {
 
     // webpackConfig.injectPlugins(new webpack.IgnorePlugin(ignoreRegex, /react-intl\/locale-data$/));
     // webpackConfig.injectPlugins(new webpack.IgnorePlugin(ignoreRegex, /intl\/locale-data\/jsonp$/));
-    webpackConfig.injectPlugins(new webpack.IgnorePlugin(ignoreRegex, /@formatjs\/intl-relativetimeformat\/locale-data$/));
-    webpackConfig.injectPlugins(new webpack.IgnorePlugin(ignoreRegex, /@formatjs\/intl-numberformat\/locale-data$/));
-    webpackConfig.injectPlugins(new webpack.IgnorePlugin(ignoreRegex, /@formatjs\/intl-listformat\/locale-data$/));
-    webpackConfig.injectPlugins(new webpack.IgnorePlugin(ignoreRegex, /@formatjs\/intl-pluralrules\/locale-data$/));
+    webpackConfig.injectPlugins(new webpack.IgnorePlugin({ resourceRegExp: ignoreRegex, contextRegExp: /@formatjs\/intl-relativetimeformat\/locale-data$/ }));
+    webpackConfig.injectPlugins(new webpack.IgnorePlugin({ resourceRegExp: ignoreRegex, contextRegExp: /@formatjs\/intl-numberformat\/locale-data$/ }));
+    webpackConfig.injectPlugins(new webpack.IgnorePlugin({ resourceRegExp: ignoreRegex, contextRegExp: /@formatjs\/intl-listformat\/locale-data$/ }));
+    webpackConfig.injectPlugins(new webpack.IgnorePlugin({ resourceRegExp: ignoreRegex, contextRegExp: /@formatjs\/intl-pluralrules\/locale-data$/ }));
   }
 }